Add basic Stubby Manager helper program.
This idea is that this small program is setuid root, and so has permissions to launch, stop the Stubby daemon and so on. It is called by the Stubby Manager, which therefore does not have to be root. In the future the manager will ask for authorisation and pass an auth token to this helper, so the helper can't be run by any old user. But that tightening of security is for the future.