Include getdns as submodule of stubby, and build if checked out.
To allow a one-stop build for stubby and getdns, include getdns as a submodule of stubby.

If the getdns submodule is checked out, build it and use that build when building and linking stubby. Note that this requires introducing libtool into the stubby build process :-(.

To match the corresponding change removing stubby and libyaml from getdns, require libyaml to build stubby.