Remove all unnecessary quotes from the example config file
YAML does not require strings to be quoted, unless they have special
characters. In the example stubby config, some strings are not quoted
but others are. In one case, the quoting has also introduced a stray
space. It is better to just write strings unquoted. This commit
removes all unneeded quotes from the file.