Rework system and user config directories on Windows.
On Unix systems, Stubby looks for a user configuration file at $HOME/.stubby.yml, and if that isn't found it looks for a system configuration file at /etc/stubby/stubby.yml. These paths are hardcoded at build time.

This approach doesn't really work on Windows. Our install package installs stubby.exe and the system-wide stubby.yaml into a Stubby subdirectory in "Program Files". This is usually C:\Program Files, but it can be on another drive. And if the 32bit install is used on a 64bit Windows install, then the files go into C:\Program Files (x86)\, but if the 32bit build is used on a 32bit Windows install, the files are in C:\Program Files. So we can't hardcode anything at runtime, and on Window must do a runtime lookup.

Also, there is no HOME environment in Windows, so the user configuration location fails at present. The Windows approach is to put it under the user AppData directory. GetDNS already does this, writing data to AppDate/getdns. So follow this too, using AppData/Stubby/stubby.yaml as the user configuration file location.

So change config file process so that on Windows the appropriate Program Files and AppData directory paths are looked up at runtime.