Support systemd readiness notification at startup
systemd services cannot reliably depend on stubby through systemd
dependencies currently. This is because systemd only knows when the
stubby daemon has started, not when it is ready to serve. In certain
situations with heavy load at startup, there can be a gap of up to
several seconds between "process started" and "sockets listened on".
During that gap, DNS queries are dropped, causing dependent services to
fail their own startup.

To avoid this issue, add support for systemd's sd_notify functionality.
This allows reporting precisely to systemd when the daemon is ready to
serve.

The current logic for when stubby is ready is hacky: we assume that we
are ready to serve before the event loop is started, when it would be
better to do so after the event loop has been started. getdns does not
currently provide any way to register callbacks on the event loop to
report readiness.

Fixes #196