Move message resources into separate DLL.
Having the messages in the executable seems to keep the executable
locked, even when you uninstall the service. This means you can't update
it without rebooting first.

Moveing the messages into a separate DLL removes this restriction. After
uninstalling the service, both the executable and the DLL can be
overwritten quite happily.