In YAML-JSON conversion, check key to see if value should be quoted.
YAML permits quotes to be omitted on string quantities. getdns
not-quite-JSON doesn't.

So use keys to detemine if values should be quoted in JSON, rather than
rely on whether they were quoted in the YAML. This does require
maintaining a list of keys that do require quoted values, and is very
simplistic - it only works reliably when the keys in question do not
have sub-keys. But it's sufficient for the Stubby config as it is at
present.