Allow PID file to be disabled at compile-time
Stubby unconditionally attempts to write a PID file under the run state
directory when it starts, but this directory is not always writeable by
non-root or non-system account users (e.g. on Slackware). To aid with
running Stubby in environments where the PID file would otherwise be
unwriteable or unnecessary, allow it to be disabled at compile-time with
the `ENABLE_PID_FILE` option (which remains on by default).