Log to syslog when running in background
Log messages are unconditionally written to stderr, and are therefore
lost after the fork when running in the background (`-g`). If syslog
support is detected at compile time, and provided we're not on a
platform for which an alternative non-stderr logging facility is
provided (i.e. Windows), send log messages to syslog when running in the
background.